+++ src/proximity/index.tsx
@@ -1,63 +1,21 @@
-import React from 'react'
-import cc from 'classcat'
-import { color } from '_utils/branding'
-import ProximityIcon from 'icon/proximityIcon'
-import style from './style'
+import styled from 'styled-components'
+import { space } from '_utils/branding'
 
-export enum Distances {
-  NONE = 'NONE',
-  CLOSE = 'CLOSE',
-  MIDDLE = 'MIDDLE',
-  FAR = 'FAR',
-}
+import Proximity from './Proximity'
 
-const size = '20px'
-
-const getColorAndTitle = (index: string, value: string, title: string) => {
-  const defaultParams = { title: '', iconColor: color.disabled }
-  switch (value) {
-    case Distances.CLOSE:
-      return index === value
-        ? {
-            title: title ? title : defaultParams.title,
-            iconColor: color.proximityClose,
-          }
-        : defaultParams
-
-    case Distances.MIDDLE:
-      return index === value
-        ? {
-            title: title ? title : defaultParams.title,
-            iconColor: color.proximityMiddle,
-          }
-        : defaultParams
-
-    case Distances.FAR:
-      return index === value
-        ? {
-            title: title ? title : defaultParams.title,
-            iconColor: color.proximityFar,
-          }
-        : defaultParams
-
-    default:
-      return defaultParams
+const StyledProximity = styled(Proximity)`
+  & {
+    display: inline-block;
+    line-height: 0;
   }
-}
 
-interface ProximityProps {
-  readonly className?: Classcat.Class
-  readonly value: Distances
-  readonly title?: string
-}
+  & svg {
+    margin-top: ${space.s};
+  }
 
-const Proximity = ({ value, title, className }: ProximityProps) => (
-  <div className={cc(['kirk-proximity', className])}>
-    {[Distances.CLOSE, Distances.MIDDLE, Distances.FAR].map(distance => (
-      <ProximityIcon key={distance} size={size} {...getColorAndTitle(distance, value, title)} />
-    ))}
-    <style jsx>{style}</style>
-  </div>
-)
-
-export default Proximity
+  & svg + svg {
+    margin-left: ${space.s};
+  }
+`
+export { Distances } from './Proximity'
+export default StyledProximity
